+++ reusables/wrappers.py
@@ -170,7 +170,7 @@ def time_it(log=None, message=None, append=None):
                 msg = _add_args(msg, *args, **kwargs)
 
             time_func = (time.perf_counter if python_version >= (3, 3)
-                         else time.clock)
+                         else time.time)
             start_time = time_func()
             try:
                 return func(*args, **kwargs)
@@ -277,20 +277,33 @@ def log_exception(log="reusables", message=None,
     return func_wrapper
 
 
-def catch_it(exceptions=(Exception, ), default=None):
+def catch_it(exceptions=(Exception, ), default=None, handler=None):
     """
     If the function encounters an exception, catch it, and
-    return the specified default instead
+    return the specified default or sent to a handler function instead.
+
+    .. code :: python
+
+        def handle_error(exception, func, *args, **kwargs):
+            print(f"{func.__name__} raised {exception} when called with {args}")
+
+        @reusables.catch_it(handler=err_func)
+        def will_raise(message="Hello")
+            raise Exception(message)
+
 
     :param exceptions: tuple of exceptions to catch
     :param default: what to return if the exception is caught
+    :param handler: function to send exception, func, *args and **kwargs
     """
     def func_wrapper(func):
         @wraps(func)
         def wrapper(*args, **kwargs):
             try:
                 return func(*args, **kwargs)
-            except exceptions:
+            except exceptions as err:
+                if handler:
+                    return handler(err, func, *args, **kwargs)
                 return default
         return wrapper
     return func_wrapper